feat(site): Add new filter to users
